Изменяет сортировку постов
