+++ js/main.js
@@ -1,7 +1,8 @@
-import './render-posts.js';
+import './render-sorted-posts.js';
 import './actions-modal.js';
 import './form.js';
 import './form-validate.js';
 import './form-message.js';
-
+import { renderSortedPosts } from './render-sorted-posts.js';
+renderSortedPosts();
