+++ js/render-sorted-posts.js
@@ -1,35 +1,21 @@
 // Сортировка миниатюр фотографий других пользователей
 
-// Массив с сервера
-import {data} from './api.js';
+import { createRandomPosts, showAlert } from './util.js';
+import { data } from './api.js';
 import { createMiniaturePosts } from './create-miniature-posts.js';
-import { showAlert, generateArrayUniqueNumbers} from './util.js';
+
 // Секция с фильтрами
 const imageFilters = document.querySelector('.img-filters');
 // По умолчанию
 const imageFilterDefault = document.querySelector('#filter-default');
-// Случайные
-const imageFilterRandom = document.querySelector('#filter-random');
-// Обсуждаемые
-const imageFilterDiscussed = document.querySelector('#filter-discussed');
-// Делаем копию массива с сервера
-const copyPosts = data.slice();
 // Сортировка по умолчанию(при открытии страницы)
 let currentFilter = imageFilterDefault.id;
 
-// <По умолчанию — фотографии в изначальном порядке с сервера - передаем в main полученный с сервера массив data>
+// Делаем копию массива с сервера
+const copyPosts = data.slice();
+
+// <По умолчанию — фотографии в изначальном порядке с сервера(data)
 // <Случайные — 10 случайных, не повторяющихся фотографий>
-// Генерируем 10 неповторяющихся чисел от 1 до 25, т.к. в данном случае массив с сервера содержит только 25 постов
-const randomNumbers = generateArrayUniqueNumbers(1, 25);
-const createRandomPosts = () => {
-  const randomPosts = [];
-  for (let i = 0; i < randomNumbers.length; i++) {
-    const posts = data.find((post) => randomNumbers[i] === post.id);
-    randomPosts.push(posts);
-  }
-  return randomPosts;
-};
-// Показываем 10 случайных неповторяющихся постов
 const randomData = createRandomPosts();
 
 // <Обсуждаемые — фотографии, отсортированные в порядке убывания количества комментариев>
@@ -40,24 +26,15 @@ const discussedData = copyPosts.sort(comparePosts);
 
 // Объект с вариантами сортировки постов
 const SortOption = {
-  'filter-default': {
-    array: data,
-    button: imageFilterDefault,
-  },
-  'filter-random': {
-    array: randomData,
-    button: imageFilterRandom,
-  },
-  'filter-discussed': {
-    array: discussedData,
-    button: imageFilterDiscussed,
-  },
+  'filter-default': data,
+  'filter-random': randomData,
+  'filter-discussed': discussedData,
 };
 
-const renderPosts = () => {
-  const array = SortOption[currentFilter].array;
+const renderPosts = async () => {
+  const array = SortOption[currentFilter];
   try {
-    createMiniaturePosts(array);
+    await createMiniaturePosts(array);
   } catch (err) {
     showAlert(err.message);
   }
@@ -77,4 +54,4 @@ const renderSortedPosts = () => {
   });
 };
 
-renderSortedPosts();
+export {renderSortedPosts};
